pyproject.toml
[project]
name = "cia-app-back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "bcrypt>=4.3.0",
    "fastapi>=0.116.1",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
